Improved boundary checking.
The game ends if the player crosses any of the four borders.  There is an issue with the top and bottom border in which it appears to print a blank line immediately before the border on endgame.  This is a strictly visual problem that will need to be resolved later; however, the logic is functioning correctly.