Added Game class
The Game class will handle user input and player interactions.  It currently includes only scripts for testing object instantiation and player movement.

The user can currently move around the board using the w-a-s-d keys, but since the input method is blocking, the player waits for input; the game is not in real time yet.  It also waits for Enter following input as raw! input mode causes the board to be drawn incorrectly.