Game loop case statement changes direction parameter
The game loop now moves the player each iteration based on their direction variable.  If the user provides input, it change the direction accordinginly.

The sleep value of each iteration is based on the player's speed.  This duration decreases each time the player eats a bittle, increasing the pace of the game.