Add score attribute to Player
The player earns points for eating bittles.  The Game class is responsible for calculating points at endgame.  Base points are based on tail length.  A bonus is awarded for playing with many abstacles since this increases difficulty.  A penalty is issued based on board size since larger boards reduce difficulty.