Add sound effects
The game now plays a welcome message at the beginning and reads out the game over message after displaying it.  A sound effect was added for when the player eats a bittle; however, this caused a brief delay as the game paused while the sound effect played.  As such, it was removed pending further research.